fix: move EnvironmentBanner to the Navigation
