+++ src/BrandedNavBar/index.ts
@@ -1,8 +1,6 @@
 export { default as BrandedNavBar } from "./NavBar";
 export { default as MenuTrigger } from "./MenuTrigger";
 export type { MenuTriggerProps } from "./MenuTrigger";
-export { default as EnvironmentBanner } from "./EnvironmentBanner";
-export type { EnvironmentBannerProps } from "./EnvironmentBanner";
 export { default as NavBarBackground } from "./NavBarBackground";
 export type { NavBarBackgroundProps } from "./NavBarBackground";
 export { default as BrandLogoContainer } from "./BrandLogoContainer";
