+++ src/Layout/ApplicationFrame.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { ZIndexProps } from "styled-system";
 import Box from "../Box/Box";
 import Flex, { FlexProps } from "../Flex/Flex";
-import EnvironmentBanner from "../BrandedNavBar/EnvironmentBanner";
+import EnvironmentBanner from "../Navigation/components/EnvironmentBanner/EnvironmentBanner";
 
 type ApplicationFrameProps = FlexProps & {
   navBar?: React.ReactNode;
