+++ src/Navigation/components/EnvironmentBanner/EnvironmentBanner.tsx
@@ -1,8 +1,7 @@
 import React from "react";
-import { Box } from "../Box";
-import { Text } from "../Type";
+import { Box } from "../../../Box";
+import { Text } from "../../../Type";
 
-/** @deprecated The BrandedNavBar component is deprecated. Use the Navigation component instead. */
 export type EnvironmentBannerProps = {
   children?: React.ReactNode;
 };
@@ -15,5 +14,4 @@ const EnvironmentBanner = ({ children }: EnvironmentBannerProps) => (
   </Box>
 );
 
-/** @deprecated The BrandedNavBar component is deprecated. Use the Navigation component instead. */
 export default EnvironmentBanner;
