+++ src/Navigation/components/EnvironmentBanner/index.ts
@@ -1,0 +1,2 @@
+export { default as EnvironmentBanner } from "./EnvironmentBanner";
+export type { EnvironmentBannerProps } from "./EnvironmentBanner";
