+++ src/Navigation/index.ts
@@ -2,6 +2,9 @@ export { default as Navigation } from "./Navigation";
 export { NulogyLogo } from "./components/NulogyLogo/NulogyLogo";
 export { NavigationLogoLink } from "./components/shared/NavigationLogoLink";
 export { NavigationLogo } from "./components/shared/NavigationLogo";
+export { EnvironmentBanner } from "./components/EnvironmentBanner";
+
 export type { NavigationProps } from "./Navigation";
 export type { MenuItems, MenuItem, UserMenu, UserMenuItem, UserMenuInfo } from "./types";
 export type { AppSwitcherConfig } from "./components/AppSwitcher/NulogyAppSwitcher";
+export type { EnvironmentBannerProps } from "./components/EnvironmentBanner";
